This commit fixes various compiler warnings on the example code.
1. "NULL" is not a suitable value for a char. Assigning NULL to a char
(or returning NULL from a char typed function) gets you a warning and
will ultimately result in a tuncated zero value, the same value you would
get if you used the 0 char value '\0';

2. The compiler will give you a warning if don't give functions a return type.
If you don't it will assume you meant "int". This fix explictly uses "int" to
resolve the warning. This is the right thing to do for instances of the "main"
function that lacked a type. I also did it for the "harness" function for
consistency, but you could argue, and I perhaps would, that "void" would
have been more appropriate.

3. There are various files that #included stdlib.h, but only so they could
use NULL, but since NULL should not have been used the #include can be
removed.